requirements.txt
blinker==1.9.0
cachecontrol==0.14.3
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-Mail==0.10.0
fpdf==1.7.2
google-auth==2.40.1
google-auth-oauthlib==1.2.2
google-auth-httplib2
google-api-python-client
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
playwright==1.55.0
psycopg2-binary==2.9.11
pyee==13.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
six==1.17.0
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
xlsxwriter==3.2.9
Flask-SQLAlchemy
Flask-Migrate